+++ bukkit/src/main/java/com/github/games647/fastlogin/bukkit/DefaultPasswordGenerator.java
@@ -1,0 +1,12 @@
+package com.github.games647.fastlogin.bukkit;
+
+import org.apache.commons.lang.RandomStringUtils;
+import org.bukkit.entity.Player;
+
+public class DefaultPasswordGenerator implements PasswordGenerator {
+
+    @Override
+    public String getRandomPassword(Player player) {
+        return RandomStringUtils.random(8, true, true);
+    }
+}