+++ bukkit/src/main/java/com/github/games647/fastlogin/bukkit/ForceLoginTask.java
@@ -82,7 +82,7 @@ public class ForceLoginTask implements Runnable {
     private boolean forceRegister(BukkitAuthPlugin authPlugin, Player player) {
         plugin.getLogger().log(Level.FINE, "Register player {0}", player.getName());
 
-        String generatedPassword = plugin.generateStringPassword();
+        String generatedPassword = plugin.generateStringPassword(player);
         boolean success = authPlugin.forceRegister(player, generatedPassword);
         player.sendMessage(ChatColor.DARK_GREEN + "Auto registered with password: " + generatedPassword);
         player.sendMessage(ChatColor.DARK_GREEN + "You may want change it?");