+++ bukkit/src/main/java/com/github/games647/fastlogin/bukkit/PasswordGenerator.java
@@ -1,0 +1,8 @@
+package com.github.games647.fastlogin.bukkit;
+
+import org.bukkit.entity.Player;
+
+public interface PasswordGenerator {
+
+    String getRandomPassword(Player player);
+}